Move files around to make things easier to extend.
